Update readme and remove example readme (#23)
* Rewrite Readme.md

Incorporate the example readme as well as additional information that
might be useful for beginners

* Delete example Readme.md

Not needed anymore as everything is in the
main Readme (consolidation of information)

* Add explanation for setting credentials via .tfvars file

* Explain duality of credentials in readme

* Rewrite authentication paragraph

Instead of explaining different authentication methods,
a blog post is referenced. The importance of explicitly
setting auth_url, username and password is highlighted.